requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
bcrypt==4.1.3
PyJWT==2.8.0
stripe==9.9.0
requests==2.31.0
python-multipart==0.0.9
